侧边栏 & 返回数据统一处理

+++ src/main/java/com/login/user_login/domain/Sidebar.java
@@ -18,7 +18,7 @@ public class Sidebar {
      * 侧边栏id
      */
     @TableId
-    private Byte sidebarId;
+    private Long sidebarId;
     /**
      * 侧边栏路径
      */
